# CityDiscoverAPI/CityDiscoverAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CityDiscoverAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Autenticacion": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsAutenticacionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Autenticacion/Validar": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsAutenticacionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FotoAnexada": {
      "get": {
        "tags": [
          "FotoAnexada"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsFotoAnexadaDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FotoAnexada"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsFotoAnexadaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "FotoAnexada"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsFotoAnexadaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FotoAnexada/{id}": {
      "get": {
        "tags": [
          "FotoAnexada"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsFotoAnexadaDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FotoAnexada"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rol": {
      "get": {
        "tags": [
          "Rol"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsRolDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rol"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsRolDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Rol"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsRolDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rol/{id}": {
      "get": {
        "tags": [
          "Rol"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsRolDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rol"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TipoLugar": {
      "get": {
        "tags": [
          "TipoLugar"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsTipoLugarDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TipoLugar"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsTipoLugarDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "TipoLugar"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsTipoLugarDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TipoLugar/{id}": {
      "get": {
        "tags": [
          "TipoLugar"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsTipoLugarDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TipoLugar"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Ubicacion": {
      "get": {
        "tags": [
          "Ubicacion"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsUbicacionDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ubicacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsUbicacionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Ubicacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/clsUbicacionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Ubicacion/{id}": {
      "get": {
        "tags": [
          "Ubicacion"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/clsUbicacionDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ubicacion"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "clsAutenticacionDTO": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
    "clsFotoAnexadaDTO": {
      "type": "object",
      "properties": {
        "idFotoAnexada": {
          "format": "int32",
          "type": "integer"
        },
        "idUbicacion": {
          "format": "int32",
          "type": "integer"
        },
        "bnImagen": {
          "format": "byte",
          "type": "string"
        }
      }
    },
    "clsRolDTO": {
      "type": "object",
      "properties": {
        "idRol": {
          "format": "int32",
          "type": "integer"
        },
        "strNombre": {
          "type": "string"
        }
      }
    },
    "clsTipoLugarDTO": {
      "type": "object",
      "properties": {
        "idTipoLugar": {
          "format": "int32",
          "type": "integer"
        },
        "strNombre": {
          "type": "string"
        }
      }
    },
    "clsUbicacionDTO": {
      "type": "object",
      "properties": {
        "idUbicacion": {
          "format": "int32",
          "type": "integer"
        },
        "numLatitud": {
          "format": "double",
          "type": "number"
        },
        "numLongitud": {
          "format": "double",
          "type": "number"
        },
        "strNombre": {
          "type": "string"
        },
        "strDescripcion": {
          "type": "string"
        },
        "idTipoLugar": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Token"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}