# CityDiscoverAPI/CityDiscoverDTO/DTOs/clsFotoAnexadaDTO.cs
namespace CityDiscoverDTO.DTOs
{
    public class clsFotoAnexadaDTO
    {
        public int IdFotoAnexada { get; set; }
        public int IdUbicacion { get; set; }

        public byte[]? bnImagen { get; set; }
    }
}